third_party/blink/web_tests/flag-specific/disable-layout-ng/external/wpt/css/css-tables/html5-table-formatting-fixed-layout-1-expected.txt

This is a testharness.js-based test.
FAIL Redundant columns of fixed-layout tables are not being merged assert_equals: expected 100 but got 50
FAIL Redundant rows of fixed-layout tables are not being merged assert_equals: expected 50 but got 25
Harness: the test ran to completion.
